# kitty: calculate damage out properly #1618